Massively simplify request_handler.
We pass the unparsed body on to the handler function now. This removes a
dependency on the schema library, and makes it possible for APIs to do
more sophisticated validation themselves.

This also removes the unnecessarily complex marshalling of arguments;
turns out you can still pass positional arguments by name. Yay for
knowing your language.